street_incident_ai/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
awsiotsdk>=1.22.0
loguru>=0.7.2
numpy>=1.24.0
opencv-python>=4.9.0.80
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.31.0
supervision>=0.21.0
ultralytics>=8.3.0
python-vlc>=3.0.20123
mypy>=1.8.0
ruff>=0.4.0
